requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.1
aiofiles==23.2.1